+++ RELEASE_NOTES.md
@@ -29,6 +29,8 @@ Release Notes ⋮
     * New projects default to using the faster lxml parser rather than BeautifulSoup.
 
 * Serving improvements
+    * When visiting a page that hasn't been downloaded,
+      a progress bar is now shown in the browser.
     * The Source dropdown on the Not in Archive page uses the same labels
       as the Source dropdown in the New Group Dialog.
     * Ignore timeout errors when responding to a request.